requirements.txt
loguru
psutil
pyyaml
pygeoip